Move costLog to test
