+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/matcher/MatchTest.scala
@@ -1,6 +1,9 @@
 package coop.rchain.rholang.interpreter.matcher
 
+import cats._
 import cats.data._
+import cats.effect._
+import cats.effect.concurrent.Ref
 import cats.implicits._
 import cats.mtl._
 import cats.mtl.implicits._
@@ -935,6 +938,18 @@ class VarMatcherSpec extends FlatSpec with Matchers with TimeLimits with TripleE
       spatialMatch[NonDetFreeMapWithCost, Par, Par](target, pattern).runFirstWithCost(Cost(0))
     res should be(Left(OutOfPhlogistonsError))
   }
+
+  def costLog[M[_]: Sync](): FunctorListen[M, Chain[Cost]] =
+    new DefaultFunctorListen[M, Chain[Cost]] {
+      override val functor: Functor[M]  = implicitly[Functor[M]]
+      private val ref                   = Ref.unsafe(Chain.empty[Cost])
+      def tell(l: Chain[Cost]): M[Unit] = ref.modify(c => (c.concat(l), ()))
+      def listen[A](fa: M[A]): M[(A, Chain[Cost])] =
+        for {
+          a <- fa
+          r <- ref.get
+        } yield ((a, r))
+    }
 
   private def doMatchAndCharge(initialPhlo: Int) = {
     implicit val costL = costLog[Task]
